requirements.txt
fastapi
uvicorn
jinja2
python-multipart
weasyprint
pytz
python-dotenv
supabase


